split cab signal tests in unit and integration tests
