+++ java/test/jmri/jmrix/can/cbus/CbusCabSignalIT.java
@@ -1,7 +1,5 @@
 package jmri.jmrix.can.cbus;
 
-import jmri.Block;
-import jmri.BlockManager;
 import jmri.DccLocoAddress;
 import jmri.InstanceManager;
 import jmri.jmrix.can.CanSystemConnectionMemo;
@@ -9,44 +7,18 @@ import jmri.jmrix.can.TrafficController;
 import jmri.jmrix.can.TrafficControllerScaffold;
 import jmri.util.JUnitUtil;
 import org.junit.After;
-import org.junit.Assert;
 import org.junit.Before;
-import org.junit.Test;
 
 /**
- * Unit Tests for CBus Cab Signals
+ * Integration Tests for CBUS Cab Signals
  *
  * @author Paul Bender Copyright (C) 2017
  * @author Steve Young Copyright (C) 2019
  */
-public class CbusCabSignalTest extends jmri.implementation.DefaultCabSignalTest {
+public class CbusCabSignalIT extends jmri.implementation.DefaultCabSignalIT {
 
     private CanSystemConnectionMemo memo;
     private TrafficController tc;
-
-    @Test
-    @Override
-    public void testSetBlock() {
-        CbusCabSignal acs = new CbusCabSignal(memo,new DccLocoAddress(1234,true)){
-            @Override
-            public jmri.SignalMast getNextMast(){
-                // don't check for signal masts, they aren't setup for this
-                // test.
-                return null;
-            }
-        };
-
-        Block b1 = InstanceManager.getDefault(BlockManager.class).provideBlock("IB12");
-        // set the block contents to our locomotive address.
-        b1.setValue(new DccLocoAddress(1234,true));
-        // call setBlock() for the cab signal.
-        acs.setBlock();
-        // and verify getBlock returns the block we set.
-        Assert.assertEquals("Block set",b1,acs.getBlock());
-
-        acs.dispose(); // verify no exceptions
-    }
-
 
     // The minimal setup for log4J
     @Before
@@ -55,9 +27,14 @@ public class CbusCabSignalTest extends jmri.implementation.DefaultCabSignalTest 
         JUnitUtil.setUp();
         JUnitUtil.resetProfileManager();
         JUnitUtil.initConfigureManager();
+        JUnitUtil.initInternalTurnoutManager();
+        JUnitUtil.initInternalSensorManager();
         JUnitUtil.initMemoryManager();
         InstanceManager.setDefault(jmri.BlockManager.class,new jmri.BlockManager());
         JUnitUtil.initLayoutBlockManager();
+        JUnitUtil.initDefaultSignalMastManager();
+        JUnitUtil.initSignalMastLogicManager();
+        InstanceManager.setDefault(jmri.jmrit.display.PanelMenu.class,new jmri.jmrit.display.PanelMenu());
 
         // prepare the cab signal
         memo = new CanSystemConnectionMemo();
@@ -77,6 +54,7 @@ public class CbusCabSignalTest extends jmri.implementation.DefaultCabSignalTest 
         cs.dispose();
         cs = null;
         JUnitUtil.tearDown();
+
     }
 
     // private final static Logger log = LoggerFactory.getLogger(CbusCabSignalTest.class);
